refactor(static-image): extract format detection and rendering backends
