+++ src/app/overlays/images/mod.rs
@@ -1,26 +1,19 @@
 mod cache;
+mod format;
 mod preload;
+mod prepare;
 mod present;
+mod render;
 mod state;
 mod types;
 
+use self::format::read_raster_dimensions;
 use super::super::*;
-use super::inline_image::{RenderedImageDimensions, TerminalWindowSize, read_png_dimensions};
-use image::{DynamicImage, GenericImageView, ImageFormat, ImageReader, imageops::FilterType};
-use quick_xml::{Reader, events::Event};
+use super::inline_image::TerminalWindowSize;
+use super::inline_image::read_png_dimensions;
 use ratatui::layout::Rect;
-use std::{
-    collections::hash_map::DefaultHasher,
-    env, fs,
-    fs::File,
-    hash::{Hash, Hasher},
-    io::Read,
-    path::{Path, PathBuf},
-    process::{Command, Stdio},
-    sync::Arc,
-    time::SystemTime,
-};
 
+pub(crate) use self::prepare::prepare_static_image_asset;
 pub(in crate::app) use self::types::{
     ImagePreviewState, PreparedStaticImage, PreparedStaticImageAsset, StaticImageKey,
     StaticImageOverlayMode, StaticImageOverlayPreparation, StaticImageOverlayRequest,
@@ -35,6 +28,34 @@ const STATIC_IMAGE_RENDER_CACHE_VERSION: usize = 3;
 const FAST_FORCE_RENDER_FFMPEG_RASTER_ARGS: [&str; 4] =
     ["-compression_level", "1", "-sws_flags", "fast_bilinear"];
 const DEFAULT_FFMPEG_RASTER_ARGS: [&str; 0] = [];
+
+pub(in crate::app) fn static_image_detail_label(entry: &Entry) -> Option<&'static str> {
+    format::static_image_detail_label(entry)
+}
+
+pub(in crate::app) fn ffmpeg_raster_render_args(
+    force_render_to_cache: bool,
+) -> &'static [&'static str] {
+    if force_render_to_cache {
+        &FAST_FORCE_RENDER_FFMPEG_RASTER_ARGS
+    } else {
+        &DEFAULT_FFMPEG_RASTER_ARGS
+    }
+}
+
+pub(in crate::app) fn image_target_width_px(
+    area: Rect,
+    window_size: Option<TerminalWindowSize>,
+) -> u32 {
+    render::image_target_width_px(area, window_size)
+}
+
+pub(in crate::app) fn image_target_height_px(
+    area: Rect,
+    window_size: Option<TerminalWindowSize>,
+) -> u32 {
+    render::image_target_height_px(area, window_size)
+}
 
 impl App {
     pub(in crate::app) fn prepared_static_image_for_overlay(
@@ -85,732 +106,21 @@ impl App {
     }
 }
 
-#[derive(Clone, Copy, Debug, Eq, PartialEq)]
-pub(super) enum StaticImageFormat {
-    Png,
-    Jpeg,
-    Gif,
-    Webp,
-    Svg,
-}
-
-impl StaticImageFormat {
-    fn detail_label(self) -> &'static str {
-        match self {
-            Self::Png => "PNG image",
-            Self::Jpeg => "JPEG image",
-            Self::Gif => "GIF image",
-            Self::Webp => "WebP image",
-            Self::Svg => "SVG image",
-        }
-    }
-
-    fn from_label(label: &'static str) -> Option<Self> {
-        match label {
-            "PNG image" => Some(Self::Png),
-            "JPEG image" => Some(Self::Jpeg),
-            "GIF image" => Some(Self::Gif),
-            "WebP image" => Some(Self::Webp),
-            "SVG image" => Some(Self::Svg),
-            _ => None,
-        }
-    }
-}
-
-pub(in crate::app) fn static_image_detail_label(entry: &Entry) -> Option<&'static str> {
-    static_image_format_for_entry(entry).map(StaticImageFormat::detail_label)
-}
-
-pub(crate) fn prepare_static_image_asset<F>(
-    request: &jobs::ImagePrepareRequest,
-    canceled: F,
-) -> Option<PreparedStaticImageAsset>
-where
-    F: Fn() -> bool,
-{
-    if canceled() {
-        return None;
-    }
-    let format = static_image_format_for_prepare_request(request)?;
-    let source_dimensions = if format == StaticImageFormat::Svg {
-        read_svg_dimensions(&request.path)?
-    } else {
-        read_raster_dimensions(&request.path)?
-    };
-    if canceled() {
-        return None;
-    }
-    let target_width_px = request.target_width_px.max(1);
-    let target_height_px = request.target_height_px.max(1);
-    let key = StaticImageKey::from_parts(
-        request.path.clone(),
-        request.size,
-        request.modified,
-        target_width_px,
-        target_height_px,
-        request.force_render_to_cache,
-        request.prepare_inline_payload,
-    );
-    let inline_payload = |path: &Path| -> Option<Option<Arc<str>>> {
-        if !request.prepare_inline_payload {
-            return Some(None);
-        }
-        Some(Some(super::inline_image::encode_iterm_inline_payload(
-            path,
-        )?))
-    };
-
-    if static_image_supports_iterm_source_passthrough_for_prepare(request, format) {
-        return Some(PreparedStaticImageAsset {
-            dimensions: source_dimensions,
-            display_path: request.path.clone(),
-            inline_payload: inline_payload(&request.path)?,
-        });
-    }
-
-    if format == StaticImageFormat::Svg {
-        let cache_path = static_image_render_cache_path(&key)?;
-        if cache_path.exists() {
-            let payload = inline_payload(&cache_path)?;
-            return Some(PreparedStaticImageAsset {
-                dimensions: prepared_display_dimensions(&cache_path, source_dimensions),
-                display_path: cache_path,
-                inline_payload: payload,
-            });
-        }
-        let temp_path = static_image_render_temp_path(&cache_path)?;
-        let rendered = (request.resvg_available
-            && render_svg_to_png_with_resvg(
-                &request.path,
-                &temp_path,
-                source_dimensions,
-                target_width_px,
-                target_height_px,
-                &canceled,
-            ))
-            || (request.magick_available
-                && render_svg_to_png_with_magick(
-                    &request.path,
-                    &temp_path,
-                    target_width_px,
-                    target_height_px,
-                    &canceled,
-                ));
-        if rendered {
-            finalize_static_image_render(&temp_path, &cache_path)?;
-            let payload = inline_payload(&cache_path)?;
-            return Some(PreparedStaticImageAsset {
-                dimensions: prepared_display_dimensions(&cache_path, source_dimensions),
-                display_path: cache_path,
-                inline_payload: payload,
-            });
-        }
-        let _ = fs::remove_file(temp_path);
-        return None;
-    }
-
-    let cache_path = static_image_render_cache_path(&key)?;
-    if cache_path.exists() {
-        let payload = inline_payload(&cache_path)?;
-        return Some(PreparedStaticImageAsset {
-            dimensions: prepared_display_dimensions(&cache_path, source_dimensions),
-            display_path: cache_path,
-            inline_payload: payload,
-        });
-    }
-    if canceled() {
-        return None;
-    }
-    let temp_path = static_image_render_temp_path(&cache_path)?;
-
-    if request.ffmpeg_available
-        && should_render_raster_with_ffmpeg(format)
-        && render_raster_to_png_with_ffmpeg(
-            &request.path,
-            &temp_path,
-            target_width_px,
-            target_height_px,
-            request.force_render_to_cache,
-            &canceled,
-        )
-    {
-        finalize_static_image_render(&temp_path, &cache_path)?;
-        let payload = inline_payload(&cache_path)?;
-        return Some(PreparedStaticImageAsset {
-            dimensions: prepared_display_dimensions(&cache_path, source_dimensions),
-            display_path: cache_path,
-            inline_payload: payload,
-        });
-    }
-
-    let image = ImageReader::open(&request.path)
-        .ok()?
-        .with_guessed_format()
-        .ok()?
-        .decode()
-        .ok()?;
-    if canceled() {
-        return None;
-    }
-    let image = apply_raster_orientation(image, read_exif_orientation(&request.path).unwrap_or(1));
-    if canceled() {
-        return None;
-    }
-    let image = shrink_image_to_fit(image, target_width_px, target_height_px);
-    if canceled() {
-        return None;
-    }
-    image.save_with_format(&temp_path, ImageFormat::Png).ok()?;
-    finalize_static_image_render(&temp_path, &cache_path)?;
-    let payload = inline_payload(&cache_path)?;
-
-    Some(PreparedStaticImageAsset {
-        dimensions: prepared_display_dimensions(&cache_path, source_dimensions),
-        display_path: cache_path,
-        inline_payload: payload,
-    })
-}
-
-fn prepared_display_dimensions(
-    display_path: &Path,
-    fallback: RenderedImageDimensions,
-) -> RenderedImageDimensions {
-    read_png_dimensions(display_path)
-        .or_else(|| read_raster_dimensions(display_path))
-        .unwrap_or(fallback)
-}
-
-fn static_image_render_cache_path(key: &StaticImageKey) -> Option<PathBuf> {
-    let mut hasher = DefaultHasher::new();
-    STATIC_IMAGE_RENDER_CACHE_VERSION.hash(&mut hasher);
-    key.path.hash(&mut hasher);
-    key.size.hash(&mut hasher);
-    key.modified.hash(&mut hasher);
-    key.target_width_px.hash(&mut hasher);
-    key.target_height_px.hash(&mut hasher);
-    key.force_render_to_cache.hash(&mut hasher);
-    let cache_dir = env::temp_dir().join(format!(
-        "elio-image-preview-v{STATIC_IMAGE_RENDER_CACHE_VERSION}"
-    ));
-    fs::create_dir_all(&cache_dir).ok()?;
-    Some(cache_dir.join(format!("image-{:016x}.png", hasher.finish())))
-}
-
-fn static_image_render_temp_path(path: &Path) -> Option<PathBuf> {
-    let parent = path.parent()?;
-    fs::create_dir_all(parent).ok()?;
-    let unique = SystemTime::now()
-        .duration_since(SystemTime::UNIX_EPOCH)
-        .ok()
-        .map(|duration| duration.as_nanos())
-        .unwrap_or_default();
-    let stem = path.file_stem()?.to_string_lossy();
-    let extension = path.extension().and_then(|extension| extension.to_str());
-    let file_name = match extension {
-        Some(extension) if !extension.is_empty() => {
-            format!(".{stem}.tmp-{}-{unique}.{extension}", std::process::id())
-        }
-        _ => format!(".{stem}.tmp-{}-{unique}", std::process::id()),
-    };
-    Some(parent.join(file_name))
-}
-
-fn finalize_static_image_render(temp_path: &Path, cache_path: &Path) -> Option<()> {
-    match fs::rename(temp_path, cache_path) {
-        Ok(()) => Some(()),
-        Err(_) if cache_path.exists() => {
-            let _ = fs::remove_file(temp_path);
-            Some(())
-        }
-        Err(_) => {
-            let _ = fs::remove_file(temp_path);
-            None
-        }
-    }
-}
-
-fn static_image_can_prepare_inline(
-    size: u64,
-    format: StaticImageFormat,
-    ffmpeg_available: bool,
-) -> bool {
-    match format {
-        StaticImageFormat::Png => true,
-        StaticImageFormat::Jpeg | StaticImageFormat::Gif | StaticImageFormat::Webp => {
-            if ffmpeg_available {
-                size <= STATIC_IMAGE_INLINE_EXTERNAL_PREPARE_MAX_BYTES
-            } else {
-                size <= STATIC_IMAGE_INLINE_FALLBACK_PREPARE_MAX_BYTES
-            }
-        }
-        StaticImageFormat::Svg => false,
-    }
-}
-
-fn static_image_supports_iterm_source_passthrough(request: &StaticImageOverlayRequest) -> bool {
-    static_image_format_for_overlay_request(request)
-        .is_some_and(|format| static_image_supports_iterm_source_format(&request.path, format))
-        && !request.force_render_to_cache
-}
-
-fn static_image_supports_iterm_source_passthrough_for_prepare(
-    request: &jobs::ImagePrepareRequest,
-    format: StaticImageFormat,
-) -> bool {
-    request.prepare_inline_payload
-        && !request.force_render_to_cache
-        && static_image_supports_iterm_source_format(&request.path, format)
-}
-
-fn static_image_supports_iterm_source_format(path: &Path, format: StaticImageFormat) -> bool {
-    match format {
-        StaticImageFormat::Png => true,
-        StaticImageFormat::Jpeg => read_exif_orientation(path).unwrap_or(1) == 1,
-        StaticImageFormat::Gif | StaticImageFormat::Webp | StaticImageFormat::Svg => false,
-    }
-}
-
-fn should_render_raster_with_ffmpeg(format: StaticImageFormat) -> bool {
-    matches!(
-        format,
-        StaticImageFormat::Jpeg | StaticImageFormat::Gif | StaticImageFormat::Webp
-    )
-}
-
-fn static_image_format_for_entry(entry: &Entry) -> Option<StaticImageFormat> {
-    crate::file_info::inspect_path_cached(&entry.path, entry.kind, entry.size, entry.modified)
-        .specific_type_label
-        .and_then(StaticImageFormat::from_label)
-}
-
-fn static_image_format_for_overlay_request(
-    request: &StaticImageOverlayRequest,
-) -> Option<StaticImageFormat> {
-    crate::file_info::inspect_path_cached(
-        &request.path,
-        EntryKind::File,
-        request.size,
-        request.modified,
-    )
-    .specific_type_label
-    .and_then(StaticImageFormat::from_label)
-}
-
-fn static_image_format_for_prepare_request(
-    request: &jobs::ImagePrepareRequest,
-) -> Option<StaticImageFormat> {
-    crate::file_info::inspect_path_cached(
-        &request.path,
-        EntryKind::File,
-        request.size,
-        request.modified,
-    )
-    .specific_type_label
-    .and_then(StaticImageFormat::from_label)
-}
-
-fn static_image_format_for_path(path: &Path) -> Option<StaticImageFormat> {
-    crate::file_info::inspect_path(path, EntryKind::File)
-        .specific_type_label
-        .and_then(StaticImageFormat::from_label)
-}
-
-fn render_svg_to_png_with_resvg(
-    input_path: &Path,
-    output_path: &Path,
-    source_dimensions: RenderedImageDimensions,
-    target_width_px: u32,
-    target_height_px: u32,
-    canceled: &impl Fn() -> bool,
-) -> bool {
-    if let Some(parent) = output_path.parent()
-        && fs::create_dir_all(parent).is_err()
-    {
-        return false;
-    }
-
-    let (width_arg, height_arg) =
-        fit_svg_render_dimensions(source_dimensions, target_width_px, target_height_px);
-    let mut command = Command::new("resvg");
-    if let Some(width_px) = width_arg {
-        command.arg("--width").arg(width_px.to_string());
-    }
-    if let Some(height_px) = height_arg {
-        command.arg("--height").arg(height_px.to_string());
-    }
-    command
-        .arg(input_path)
-        .arg(output_path)
-        .stdout(Stdio::null())
-        .stderr(Stdio::null());
-
-    run_cancelable_command(&mut command, canceled)
-        .is_some_and(|status| status.success() && output_path.exists())
-}
-
-fn render_svg_to_png_with_magick(
-    input_path: &Path,
-    output_path: &Path,
-    target_width_px: u32,
-    target_height_px: u32,
-    canceled: &impl Fn() -> bool,
-) -> bool {
-    if let Some(parent) = output_path.parent()
-        && fs::create_dir_all(parent).is_err()
-    {
-        return false;
-    }
-
-    run_cancelable_command(
-        Command::new("magick")
-            .arg(input_path)
-            .arg("-resize")
-            .arg(format!(
-                "{}x{}>",
-                target_width_px.max(1),
-                target_height_px.max(1)
-            ))
-            .arg(output_path)
-            .stdout(Stdio::null())
-            .stderr(Stdio::null()),
-        canceled,
-    )
-    .is_some_and(|status| status.success() && output_path.exists())
-}
-
-fn fit_svg_render_dimensions(
-    source_dimensions: RenderedImageDimensions,
-    target_width_px: u32,
-    target_height_px: u32,
-) -> (Option<u32>, Option<u32>) {
-    let source_width = source_dimensions.width_px.max(1) as f32;
-    let source_height = source_dimensions.height_px.max(1) as f32;
-    let scale = (target_width_px.max(1) as f32 / source_width)
-        .min(target_height_px.max(1) as f32 / source_height)
-        .min(1.0);
-    if scale >= 1.0 {
-        return (None, None);
-    }
-
-    let fitted_width = (source_width * scale).round().max(1.0) as u32;
-    let fitted_height = (source_height * scale).round().max(1.0) as u32;
-    let width_ratio = target_width_px.max(1) as f32 / source_width;
-    let height_ratio = target_height_px.max(1) as f32 / source_height;
-    if width_ratio <= height_ratio {
-        (Some(fitted_width), None)
-    } else {
-        (None, Some(fitted_height))
-    }
-}
-
-fn render_raster_to_png_with_ffmpeg(
-    input_path: &Path,
-    output_path: &Path,
-    target_width_px: u32,
-    target_height_px: u32,
-    force_render_to_cache: bool,
-    canceled: &impl Fn() -> bool,
-) -> bool {
-    if let Some(parent) = output_path.parent()
-        && fs::create_dir_all(parent).is_err()
-    {
-        return false;
-    }
-
-    let mut command = Command::new("ffmpeg");
-    command
-        .arg("-v")
-        .arg("error")
-        .arg("-y")
-        .arg("-i")
-        .arg(input_path)
-        .arg("-frames:v")
-        .arg("1")
-        .arg("-vf")
-        .arg(format!(
-            "scale=w={}:h={}:force_original_aspect_ratio=decrease",
-            target_width_px.max(1),
-            target_height_px.max(1)
-        ));
-    command.args(ffmpeg_raster_render_args(force_render_to_cache));
-    command
-        .arg(output_path)
-        .stdout(Stdio::null())
-        .stderr(Stdio::null());
-
-    run_cancelable_command(&mut command, canceled)
-        .is_some_and(|status| status.success() && output_path.exists())
-}
-
-pub(in crate::app) fn ffmpeg_raster_render_args(
-    force_render_to_cache: bool,
-) -> &'static [&'static str] {
-    if force_render_to_cache {
-        &FAST_FORCE_RENDER_FFMPEG_RASTER_ARGS
-    } else {
-        &DEFAULT_FFMPEG_RASTER_ARGS
-    }
-}
-
-fn read_raster_dimensions(path: &Path) -> Option<RenderedImageDimensions> {
-    let (mut width_px, mut height_px) = ImageReader::open(path)
-        .ok()?
-        .with_guessed_format()
-        .ok()?
-        .into_dimensions()
-        .ok()?;
-    if exif_orientation_swaps_dimensions(read_exif_orientation(path).unwrap_or(1)) {
-        std::mem::swap(&mut width_px, &mut height_px);
-    }
-    (width_px > 0 && height_px > 0).then_some(RenderedImageDimensions {
-        width_px,
-        height_px,
-    })
-}
-
-fn read_exif_orientation(path: &Path) -> Option<u16> {
-    let mut file = File::open(path).ok()?;
-    let mut soi = [0_u8; 2];
-    file.read_exact(&mut soi).ok()?;
-    if soi != [0xff, 0xd8] {
-        return None;
-    }
-
-    loop {
-        let mut prefix = [0_u8; 1];
-        file.read_exact(&mut prefix).ok()?;
-        while prefix[0] != 0xff {
-            file.read_exact(&mut prefix).ok()?;
-        }
-
-        let mut marker = [0_u8; 1];
-        file.read_exact(&mut marker).ok()?;
-        while marker[0] == 0xff {
-            file.read_exact(&mut marker).ok()?;
-        }
-
-        match marker[0] {
-            0xd8 | 0x01 => continue,
-            0xd9 | 0xda => return None,
-            _ => {
-                let mut length = [0_u8; 2];
-                file.read_exact(&mut length).ok()?;
-                let payload_len = usize::from(u16::from_be_bytes(length)).checked_sub(2)?;
-                let mut payload = vec![0_u8; payload_len];
-                file.read_exact(&mut payload).ok()?;
-                if marker[0] == 0xe1 && payload.starts_with(b"Exif\0\0") {
-                    return parse_exif_orientation(&payload[6..]);
-                }
-            }
-        }
-    }
-}
-
-fn parse_exif_orientation(tiff: &[u8]) -> Option<u16> {
-    if tiff.len() < 8 {
-        return None;
-    }
-    let little_endian = match &tiff[..2] {
-        b"II" => true,
-        b"MM" => false,
-        _ => return None,
-    };
-    let read_u16 = |offset: usize| -> Option<u16> {
-        let bytes: [u8; 2] = tiff.get(offset..offset + 2)?.try_into().ok()?;
-        Some(if little_endian {
-            u16::from_le_bytes(bytes)
-        } else {
-            u16::from_be_bytes(bytes)
-        })
-    };
-    let read_u32 = |offset: usize| -> Option<u32> {
-        let bytes: [u8; 4] = tiff.get(offset..offset + 4)?.try_into().ok()?;
-        Some(if little_endian {
-            u32::from_le_bytes(bytes)
-        } else {
-            u32::from_be_bytes(bytes)
-        })
-    };
-
-    if read_u16(2)? != 42 {
-        return None;
-    }
-    let ifd_offset = read_u32(4)? as usize;
-    let entry_count = usize::from(read_u16(ifd_offset)?);
-    let mut entry_offset = ifd_offset + 2;
-    for _ in 0..entry_count {
-        let tag = read_u16(entry_offset)?;
-        let field_type = read_u16(entry_offset + 2)?;
-        let count = read_u32(entry_offset + 4)?;
-        if tag == 0x0112 && field_type == 3 && count >= 1 {
-            return read_u16(entry_offset + 8);
-        }
-        entry_offset += 12;
-    }
-    None
-}
-
-fn exif_orientation_swaps_dimensions(orientation: u16) -> bool {
-    matches!(orientation, 5..=8)
-}
-
-fn apply_raster_orientation(image: DynamicImage, orientation: u16) -> DynamicImage {
-    match orientation {
-        2 => image.fliph(),
-        3 => image.rotate180(),
-        4 => image.flipv(),
-        5 => image.rotate90().fliph(),
-        6 => image.rotate90(),
-        7 => image.rotate90().flipv(),
-        8 => image.rotate270(),
-        _ => image,
-    }
-}
-
-fn shrink_image_to_fit(
-    image: DynamicImage,
-    target_width_px: u32,
-    target_height_px: u32,
-) -> DynamicImage {
-    let (width_px, height_px) = image.dimensions();
-    if width_px <= target_width_px.max(1) && height_px <= target_height_px.max(1) {
-        image
-    } else {
-        image.resize(
-            target_width_px.max(1),
-            target_height_px.max(1),
-            FilterType::Triangle,
-        )
-    }
-}
-
-pub(in crate::app) fn image_target_width_px(
-    area: Rect,
-    window_size: Option<TerminalWindowSize>,
-) -> u32 {
-    let (cell_width_px, _) = image_cell_pixels(window_size);
-    (f32::from(area.width.max(1)) * cell_width_px)
-        .round()
-        .max(1.0) as u32
-}
-
-pub(in crate::app) fn image_target_height_px(
-    area: Rect,
-    window_size: Option<TerminalWindowSize>,
-) -> u32 {
-    let (_, cell_height_px) = image_cell_pixels(window_size);
-    (f32::from(area.height.max(1)) * cell_height_px)
-        .round()
-        .max(1.0) as u32
-}
-
-fn image_cell_pixels(window_size: Option<TerminalWindowSize>) -> (f32, f32) {
-    match window_size {
-        Some(window_size) => (
-            window_size.pixels_width as f32 / f32::from(window_size.cells_width.max(1)),
-            window_size.pixels_height as f32 / f32::from(window_size.cells_height.max(1)),
-        ),
-        None => (8.0, 16.0),
-    }
-}
-
-fn run_cancelable_command<F>(
-    command: &mut Command,
-    canceled: &F,
-) -> Option<std::process::ExitStatus>
-where
-    F: Fn() -> bool,
-{
-    let mut child = command.spawn().ok()?;
-    loop {
-        if canceled() {
-            let _ = child.kill();
-            let _ = child.wait();
-            return None;
-        }
-        if let Some(status) = child.try_wait().ok()? {
-            return Some(status);
-        }
-        std::thread::sleep(std::time::Duration::from_millis(10));
-    }
-}
-
-fn read_svg_dimensions(path: &std::path::Path) -> Option<RenderedImageDimensions> {
-    let bytes = fs::read(path).ok()?;
-    let mut reader = Reader::from_reader(bytes.as_slice());
-    reader.config_mut().trim_text(true);
-
-    let mut buffer = Vec::new();
-    loop {
-        match reader.read_event_into(&mut buffer).ok()? {
-            Event::Start(tag) | Event::Empty(tag) if tag.name().as_ref() == b"svg" => {
-                let mut width = None;
-                let mut height = None;
-                let mut view_box = None;
-                for attribute in tag.attributes().flatten() {
-                    let key = attribute.key.as_ref();
-                    let value = attribute.decode_and_unescape_value(reader.decoder()).ok()?;
-                    match key {
-                        b"width" => width = parse_svg_length_px(&value),
-                        b"height" => height = parse_svg_length_px(&value),
-                        b"viewBox" => view_box = parse_svg_view_box(&value),
-                        _ => {}
-                    }
-                }
-
-                return match (width, height, view_box) {
-                    (Some(width_px), Some(height_px), _) if width_px > 0 && height_px > 0 => {
-                        Some(RenderedImageDimensions {
-                            width_px,
-                            height_px,
-                        })
-                    }
-                    (_, _, Some((width_px, height_px))) if width_px > 0.0 && height_px > 0.0 => {
-                        Some(RenderedImageDimensions {
-                            width_px: width_px.round() as u32,
-                            height_px: height_px.round() as u32,
-                        })
-                    }
-                    _ => None,
-                };
-            }
-            Event::Eof => return None,
-            _ => {}
-        }
-        buffer.clear();
-    }
-}
-
-fn parse_svg_length_px(value: &str) -> Option<u32> {
-    let trimmed = value.trim();
-    let numeric = trimmed
-        .strip_suffix("px")
-        .unwrap_or(trimmed)
-        .trim()
-        .parse::<f32>()
-        .ok()?;
-    (numeric > 0.0).then_some(numeric.round() as u32)
-}
-
-fn parse_svg_view_box(value: &str) -> Option<(f32, f32)> {
-    let mut parts = value
-        .split(|character: char| character.is_ascii_whitespace() || character == ',')
-        .filter(|part| !part.is_empty());
-    let _min_x = parts.next()?.parse::<f32>().ok()?;
-    let _min_y = parts.next()?.parse::<f32>().ok()?;
-    let width = parts.next()?.parse::<f32>().ok()?;
-    let height = parts.next()?.parse::<f32>().ok()?;
-    Some((width, height))
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
     use crate::app::overlays::inline_image::{
-        OverlayPresentState, TerminalWindowSize, build_kitty_clear_sequence,
+        ImageProtocol, OverlayPresentState, RenderedImageDimensions, TerminalWindowSize,
+        build_kitty_clear_sequence,
     };
     use image::{DynamicImage, ImageFormat, Rgba, RgbaImage};
-    use std::time::{Duration, UNIX_EPOCH};
+    use ratatui::layout::Rect;
+    use std::{
+        fs,
+        path::{Path, PathBuf},
+        sync::Arc,
+        time::{Duration, SystemTime, UNIX_EPOCH},
+    };
 
     fn temp_root(label: &str) -> PathBuf {
         let unique = SystemTime::now()
